+++ src/check_python_versions/__init__.py
@@ -11,4 +11,4 @@ Makes sure the set of supported Python versions is consistent between
 
 """
 __author__ = 'Marius Gedminas <marius@gedmin.as>'
-__version__ = '0.12.1.dev0'
+__version__ = '0.12.1'