+++ CHANGES.rst
@@ -1,7 +1,7 @@
 Changelog
 =========
 
-0.12.1 (unreleased)
+0.12.1 (2019-05-02)
 -------------------
 
 - Improvements in Python version updating: